% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_boucle}
\alias{compter_nombre_boucle}
\title{Cette fonction renvoie le nombre de valeurs distinctes dans la colonne
\verb{Numéro de boucle}.}
\usage{
compter_nombre_boucle(trajet)
}
\arguments{
\item{trajet}{Un data.frame ou tibble contenant une colonne \verb{Numéro de boucle}.}
}
\value{
Un entier correspondant au nombre de boucles distinctes.
}
\description{
Cette fonction renvoie le nombre de valeurs distinctes dans la colonne
\verb{Numéro de boucle}.
}
