% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_trajets}
\alias{compter_nombre_trajets}
\title{Cette fonction calcule la somme de la colonne \code{Total} d'un tableau de trajets.}
\usage{
compter_nombre_trajets(trajet)
}
\arguments{
\item{trajet}{Un data.frame ou tibble contenant une colonne \code{Total}.}
}
\value{
Un entier correspondant au nombre total de trajets.
}
\description{
Cette fonction calcule la somme de la colonne \code{Total} d'un tableau de trajets.
}
