% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Cette fonction filtre les lignes d'un tableau de trajets pour ne conserver
que celles où la probabilité d'anomalie est manquante.}
\usage{
filtre_anomalie(trajet)
}
\arguments{
\item{trajet}{Un data.frame ou tibble contenant une colonne
\verb{Probabilité de présence d'anomalies}.}
}
\value{
Un tibble filtré.
}
\description{
Cette fonction filtre les lignes d'un tableau de trajets pour ne conserver
que celles où la probabilité d'anomalie est manquante.
}
