% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Cette fonction filtre les anomalies, calcule la distribution des trajets
par jour de la semaine, recode les jours en labels textuels, puis génère
un graphique en barres.}
\usage{
plot_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame ou tibble contenant au minimum les colonnes
\verb{Probabilité de présence d'anomalies}, \verb{Jour de la semaine} et \code{Total}.}
}
\value{
Un objet ggplot représentant la distribution hebdomadaire.
}
\description{
Cette fonction filtre les anomalies, calcule la distribution des trajets
par jour de la semaine, recode les jours en labels textuels, puis génère
un graphique en barres.
}
