% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Cette fonction identifie la ligne du tableau ayant la valeur maximale
dans la colonne \code{Total}.}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{Un data.frame ou tibble contenant les colonnes \code{Total},
\verb{Boucle de comptage} et \code{Jour}.}
}
\value{
Un tibble contenant le trajet maximal.
}
\description{
Cette fonction identifie la ligne du tableau ayant la valeur maximale
dans la colonne \code{Total}.
}
