% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Cette fonction compte le nombre total de trajets par jour de la semaine,
pondéré par la colonne \code{Total}.}
\usage{
calcul_distribution_semaine(trajet, filtre = TRUE)
}
\arguments{
\item{trajet}{Un data.frame ou tibble contenant les colonnes
\verb{Jour de la semaine} et \code{Total}.}

\item{filtre}{un booléen true/false qui filtre les données.}
}
\value{
Un tibble avec deux colonnes : le jour de la semaine et le nombre
total de trajets.
}
\description{
Cette fonction compte le nombre total de trajets par jour de la semaine,
pondéré par la colonne \code{Total}.
}
