% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer les données vélo selon un ou plusieurs numéros de boucle}
\usage{
filtrer_trajet(data, boucle = NULL)
}
\arguments{
\item{data}{Un data.frame contenant les données de comptage vélo (ex: df_velo)}

\item{boucle}{Un vecteur de numéros de boucle à filtrer (ex: c("880", "881")), il sera nulle par défaut et renverra le data.frame complet}
}
\value{
Un data.frame filtré contenant uniquement les boucles sélectionnées
}
\description{
Cette fonction sélectionne uniquement les lignes du dataset correspondant
à un ou plusieurs numéros de boucle.
}
